+++ src/ErrorEvent.php
@@ -10,16 +10,10 @@ namespace yii\queue;
 /**
  * Error Event.
  *
+ * @deprecated Will be removed in 3.0. Use ExecEvent instead.
+ *
  * @author Roman Zhuravlev <zhuravljov@gmail.com>
  */
 class ErrorEvent extends ExecEvent
 {
-    /**
-     * @var \Exception|\Throwable
-     */
-    public $error;
-    /**
-     * @var bool
-     */
-    public $retry;
 }