+++ src/ExecEvent.php
@@ -15,12 +15,28 @@ namespace yii\queue;
 class ExecEvent extends JobEvent
 {
     /**
-     * @var int attempt number
+     * @var int attempt number.
+     * @see Queue::EVENT_BEFORE_EXEC
+     * @see Queue::EVENT_AFTER_EXEC
+     * @see Queue::EVENT_AFTER_ERROR
      */
     public $attempt;
     /**
-     * @var mixed result of a job execution
+     * @var mixed result of a job execution in case job is done.
+     * @see Queue::EVENT_AFTER_EXEC
      * @since 2.1.1
      */
     public $result;
+    /**
+     * @var null|\Exception|\Throwable
+     * @see Queue::EVENT_AFTER_ERROR
+     * @since 2.1.1
+     */
+    public $error;
+    /**
+     * @var null|bool
+     * @see Queue::EVENT_AFTER_ERROR
+     * @since 2.1.1
+     */
+    public $retry;
 }