+++ src/cli/VerboseBehavior.php
@@ -10,7 +10,6 @@ namespace yii\queue\cli;
 use yii\base\Behavior;
 use yii\console\Controller;
 use yii\helpers\Console;
-use yii\queue\ErrorEvent;
 use yii\queue\ExecEvent;
 use yii\queue\JobInterface;
 
@@ -82,9 +81,9 @@ class VerboseBehavior extends Behavior
     }
 
     /**
-     * @param ErrorEvent $event
+     * @param ExecEvent $event
      */
-    public function afterError(ErrorEvent $event)
+    public function afterError(ExecEvent $event)
     {
         $this->command->stdout(date('Y-m-d H:i:s'), Console::FG_YELLOW);
         $this->command->stdout($this->jobTitle($event), Console::FG_GREY);