+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Yii2 Queue Extension Change Log
 - Bug #269: Handling of broken messages that are not unserialized correctly (zhuravljov)
 - Bug #267: Fixed symfony/process incompatibility (rob006)
 - Enh: Job execution result that forwarded to the event handler (zhuravljov)
+- Enh: ErrorEvent mark as deprecated (zhuravljov)
 
 2.1.0 May 24, 2018
 ------------------